map_reduce/requirements.txt
langgraph>=0.2.0
pydantic>=2.0.0
openai>=1.35.0
python-dotenv>=1.0.0

